fix: resolve C# build errors
- Core: Use ManagementClass to get WMI method parameters (fix GetMethodParameters error)
- Core: Fix EnergyDriver.SendCommand logic (cast to int, fix array copy)
- Core: Update NativeMethods to allow nullable byte arrays for DeviceIoControl
- Core: Ensure System.Management compatibility
- Core: Fully qualify HidSharp.DeviceList to verify reference

+++ LegionLoqControl.Core/Hardware/EnergyDriver.cs
@@ -47,16 +47,9 @@ namespace LegionLoqControl.Core.Hardware
                 if (!Open()) return false;
             }
 
-            // Structure reversed from LLT/Rust:
-            // Input buffer is usually just the value? Or a struct?
-            // Checking Rust implementation...
-            // Rust: DWORD (4 bytes).
-            // LLT: DeviceIoControl(IOCTL_ENERGY_BATTERY_CHARGE_MODE, ref input, ...)
-            // Input is `BatteryChargeModeInput` struct: { int Mode, int Enable }
-
             var input = new byte[8];
-            Array.Copy(BitConverter.GetBytes(featureId), 0, input, 0, 4); // Mode (3=Conservation)
-            Array.Copy(BitConverter.GetBytes(value), 0, input, 4, 4);     // Enable (1/0)
+            Array.Copy(BitConverter.GetBytes((int)featureId), 0, input, 0, 4);
+            Array.Copy(BitConverter.GetBytes(value), 0, input, 4, 4);
 
             uint bytesReturned;
             return NativeMethods.DeviceIoControl(
